requirements.txt
# Python version 3.9.2
numpy==1.19.5
matplotlib==3.4.1
h5py==3.2.1
scikit-learn==0.24.1
pandas==1.2.3
scipy==1.6.2
pyyaml==5.4.1
jupyter==1.0.0
seaborn==0.11.1
bokeh==2.4.3
numba==0.53.1
umap-learn==0.5.1
holoviews==1.14.5
astropy==4.3.1
dtreeviz==1.3.6
yellowbrick
ytree
